refactor: split agent control modules
